Fixed issue with input file parameter ignored, extended limits of possible amount of iterations, optimized two step hill climbing.

+++ anneal.hpp
@@ -5,7 +5,7 @@ namespace marxan {
     typedef struct sanneal
     {
         long int Titns;
-        long int iterations;
+        long long iterations;
         long int Tlen;
         double Tinit;    /* Initial Temperature */
         double Tcool;    /* Cooling Factor */
